splits palette to own file
