+++ lib/themes/main/palette.ts
@@ -1,0 +1,21 @@
+declare module "@mui/material/styles" {
+  interface Palette {
+    green: Palette["primary"];
+  }
+
+  interface PaletteOptions {
+    green?: PaletteOptions["primary"];
+  }
+}
+
+declare module "@mui/material/SvgIcon" {
+  interface SvgIconPropsColorOverrides {
+    green: true;
+  }
+}
+
+export const palette = {
+  green: {
+    main: "#00B712",
+  },
+};
