+++ lib/themes/main/setupTheme.ts
@@ -1,6 +1,7 @@
 "use client";
 import { Roboto } from "next/font/google";
 import { createTheme } from "@mui/material/styles";
+import { palette } from "./palette";
 
 const roboto = Roboto({
   weight: ["300", "400", "500", "700"],
@@ -8,28 +9,8 @@ const roboto = Roboto({
   display: "swap",
 });
 
-declare module "@mui/material/styles" {
-  interface Palette {
-    green: Palette["primary"];
-  }
-
-  interface PaletteOptions {
-    green?: PaletteOptions["primary"];
-  }
-}
-
-declare module "@mui/material/SvgIcon" {
-  interface SvgIconPropsColorOverrides {
-    green: true;
-  }
-}
-
 const setupTheme = createTheme({
-  palette: {
-    green: {
-      main: "#00B712",
-    },
-  },
+  palette,
   typography: {
     fontSize: 16,
     fontFamily: roboto.style.fontFamily,
